+++ src/bin/patternx.rs
@@ -2,8 +2,8 @@
 
 use std::sync::Arc;
 
-use vga::util;
-use vga::{set_vertical_display_end, SCReg, VGA};
+use vga::{set_vertical_display_end, SCReg};
+use vga::{util, VGABuilder};
 
 static PATT_TABLE: [[u8; 16]; 16] = [
     [10, 0, 10, 0, 0, 10, 0, 10, 10, 0, 10, 0, 0, 10, 0, 10],
@@ -33,7 +33,10 @@ static PATT_TABLE: [[u8; 16]; 16] = [
 ];
 
 pub fn main() -> Result<(), String> {
-    let (vga, handle) = VGA::setup(0x13, false)?;
+    let (vga, handle) = VGABuilder::new()
+        .video_mode(0x13)
+        .fullscreen(false)
+        .build()?;
 
     //enable Mode X
     let mem_mode = vga.get_sc_data(SCReg::MemoryMode);