+++ src/bin/rectx.rs
@@ -2,11 +2,14 @@
 
 use std::sync::Arc;
 
-use vga::util;
-use vga::{set_vertical_display_end, SCReg, VGA};
+use vga::{set_vertical_display_end, SCReg};
+use vga::{util, VGABuilder};
 
 fn main() -> Result<(), String> {
-    let (vga, handle) = VGA::setup(0x13, false)?;
+    let (vga, handle) = VGABuilder::new()
+        .video_mode(0x13)
+        .fullscreen(false)
+        .build()?;
 
     //enable Mode X
     let mem_mode = vga.get_sc_data(SCReg::MemoryMode);